weltenhub/requirements.txt
# =============================================================================
# Weltenhub Requirements
# =============================================================================
# Django + HTMX + Postgres stack for Hetzner deployment
# Last updated: 2026-02-04
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
Django>=5.0,<5.2
gunicorn>=21.0.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg[binary]>=3.1.0
dj-database-url>=2.1.0

# -----------------------------------------------------------------------------
# Caching & Background Tasks
# -----------------------------------------------------------------------------
redis>=5.0.0
django-redis>=5.4.0
celery>=5.3.0
django-celery-beat>=2.5.0

# -----------------------------------------------------------------------------
# REST API
# -----------------------------------------------------------------------------
djangorestframework>=3.14.0
django-filter>=23.0,<25.0
drf-spectacular>=0.27.0

# -----------------------------------------------------------------------------
# Security & Auth
# -----------------------------------------------------------------------------
django-cors-headers>=4.3.0
PyJWT>=2.8.0

# -----------------------------------------------------------------------------
# Frontend
# -----------------------------------------------------------------------------
django-htmx>=1.17.0
whitenoise>=6.6.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-decouple>=3.8
Pillow>=10.0.0
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0

# -----------------------------------------------------------------------------
# Monitoring & Logging
# -----------------------------------------------------------------------------
sentry-sdk>=1.38.0
structlog>=24.1.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
